+++ R/check_cran.R
@@ -8,6 +8,7 @@
 #' @inheritParams rcmdcheck::rcmdcheck
 #' @return A `Checklist` object.
 #' @importFrom assertthat assert_that
+#' @importFrom gert git_ahead_behind git_branch_exists git_info
 #' @importFrom httr HEAD
 #' @importFrom rmarkdown pandoc_exec
 #' @importFrom rcmdcheck rcmdcheck
@@ -40,9 +41,31 @@ check_cran <- function(x = ".", quiet = FALSE) {
       error_on = "never", quiet = quiet
     )
   )
+  main_branch <- ifelse(
+    is.na(git_info(repo = x$get_path)$head), "none",
+    ifelse(git_branch_exists("main", repo = x$get_path), "main", "master")
+  )
+  if (
+    main_branch != "none" &&
+    git_ahead_behind(upstream = main_branch, repo = x$get_path)$upstream ==
+      git_ahead_behind(upstream = main_branch, repo = x$get_path)$local &&
+    any(grepl("Insufficient package version", check_output$warnings))
+  ) { # nocov start
+    incoming <- grepl(
+      "checking CRAN incoming feasibility", check_output$warnings
+    )
+    gsub("
+
+Insufficient package version \\(submitted: .*, existing: .*\\)
+
+Days since last update: [0-9]+", "", check_output$warnings[incoming]
+    ) -> new_incoming
+    if (length(strsplit(new_incoming, "\n")[[1]]) == 2) {
+      check_output$warnings <- check_output$warnings[!incoming]
+    }
+  } # nocov end
   x$add_rcmdcheck(
-    errors = check_output$errors,
-    warnings = check_output$warnings,
+    errors = check_output$errors, warnings = check_output$warnings,
     notes = check_output$notes
   )
   return(x)