+++ R/check_description.R
@@ -111,6 +111,13 @@ from the web. More info on https://github.com/github/renaming"[
       !unchanged_repo(repo, status_before)
     ]
   )
+
+  # check if the language is set
+  desc_error <- c(
+    desc_error,
+    "Language field not set."[is.na(this_desc$get("Language"))]
+  )
+
   x$add_error(desc_error, "DESCRIPTION")
   x$add_notes(notes)
   x$add_warnings(check_authors(this_desc))