+++ R/check_filename.R
@@ -106,11 +106,11 @@ Failing folder: `%s`",
       c(
         "\\.[a-zA-Z]+ignore", "\\.Rprofile", "\\.[a-zA-Z]+\\.(json|yml)",
         "CITATION", "DESCRIPTION", "NAMESPACE", "CITATION.cff",
-        "README\\.R?md", "NEWS\\.md", # nolint: nonportable_path_linter.
+        "README\\.R?md", "NEWS\\.md",
         "CODE_OF_CONDUCT.md", "CONTRIBUTING.md", "LICENSE.md", "SUPPORT.md",
         "SECURITY.md", "FUNDING.yml",
         "Dockerfile", "docker-compose.*.yml",
-        ".*-package\\.Rd", "cran-comments.md", "WORDLIST"
+        ".*-package\\.Rd", "cran-comments.md", "WORDLIST.*"
       ),
       collapse = "|"
     )
@@ -129,6 +129,9 @@ Fails: `%s`",
     )
   )
 
+  # ignore .rda files in the data directory
+  to_ignore <- !grepl("^data/[a-z0-9_]*\\.rda$", files)
+
   extension <- gsub("(.*)\\.(.*)?", "\\2", basename(files))
   # extension exceptions
   exception <- grepl("^R(d|da|nw|md|proj)?$", extension) |
@@ -145,7 +148,7 @@ Fails: `%s`",
     problems,
     sprintf(
       "R file requires extension with upper case R.\nFails: `%s`",
-      files[grepl("^r(proj|d|da|md|nw)?$", extension)]
+      files[grepl("^r(proj|d|da|md|nw)?$", extension) & to_ignore]
     )
   )